requirements.txt
Django==4.1.1
psycopg2==2.9.3
Pillow==9.2.0
django-exiffield==3.0.0
jwt==1.3.1